Refactor DocList and DocsPage
Make it DRY

How it flows:

* DocsPage renders DocsTree.
* DocsTree renders rootDocList by new DocList(UI component).
* DocsTree has binded event which triggers callback.
  * callback executes router push.
  * callback makes subDocList by new DocList.
