+++ src/components/DocsPage.js
@@ -1,18 +1,18 @@
-import DocList from './DocList.js';
+import DocsTree from './DocsTree.js';
 
 import { request } from '../services/api.js';
 
 export default function DocsPage({ $target }) {
   const $page = document.createElement('div');
 
-  const documentList = new DocList({
+  const docsTree = new DocsTree({
     $target: $page,
     initialState: [],
   });
 
   this.render = async () => {
     const documents = await request('/documents');
-    documentList.setState(documents);
+    docsTree.setState(documents);
 
     $target.appendChild($page);
   };
